Rename Totrinnskontrolls -> Totrinnskontroll
