+++ src/frontend/komponenter/Behandling/Totrinnskontroll/Totrinnskontroll.tsx
